refactor(csv): one mapper for our own export coming back in
The round-trip test hand-wrote the export -> importer-row mapping, and the
release validation needs the same mapping. A second copy would let the two
disagree about a file while both passed, so it moves to its own module.

Its own module and not src/csv.ts: csv.ts is transpiled INTO the widgets by
@inlinets, so anything added there ships to every widget that inlines it,
and no widget needs this.
